Updates for front end season.

+++ Hoops.Core/Entities/VwRoles.cs
@@ -6,8 +6,8 @@ namespace Hoops.Core.Entities
     public partial class VwRoles
     {
         public int CompanyId { get; set; }
-        public decimal RoleId { get; set; }
-        public decimal UserId { get; set; }
+        public int RoleId { get; set; }
+        public int UserId { get; set; }
         public string ScreenName { get; set; }
         public DateTime? CreatedDate { get; set; }
         public string CreatedUser { get; set; }
